refactor: move quarter value to domain
