+++ src/domain/mod.rs
@@ -5,4 +5,5 @@ pub mod repository;
 
 pub use model::income::Income;
 pub use model::payment::Payment;
+pub use model::quarter::Quarter;
 pub use model::tax_payment::*;
