+++ src/domain/model/mod.rs
@@ -1,4 +1,5 @@
 //! Definitions of domain entities and values.
 pub mod income;
 pub mod payment;
+pub mod quarter;
 pub mod tax_payment;
